Version 0.68 2018-01-22
Issue 59. Added an unsold button to the auction page. Marks the post as unsold in the database, which is usefull for follow up of the auction.
Added sellers all posts to the compilation, not just the sold posts. Easier for the seller to se what was sold.
Each post is marked as checked in or not on the compilation.
Removed some debug print statements.

modified:   CHANGELOG
modified:   make_compilation_pdf.py
modified:   templates/auktion.html
modified:   uaf.py